+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `love.system.getOS()` now returns actual operating system name
 - `love.console` an in-game console toggled with the tilde key
+- `love.graphics.getColor()` to retrieve the active painting color
 
 ## 0.16.1 - 2018-03-12
 ### Added